Rename: Directory Change and conbine all file - 그동안 이곳저곳 흩어져 있던 파일들을 지우고 Udata폴더에 request와   response로 폴더를 나눠줘서 넣어줬습니다. - 서버에서 TODO 몇가지 수정했습니다. - Cpp 버전 11 때문에 그거에 맞게 수정했습니다.
